qnn unique sample, single pqc argument
