+++ tests/qhbm_test.py
@@ -23,6 +23,7 @@ import tensorflow_quantum as tfq
 from qhbmlib import qhbm
 from qhbmlib import ebm
 from qhbmlib import qnn
+from qhbmlib import util
 from tests import test_util
 
 # Global tolerance, set for float32.
@@ -41,13 +42,14 @@ class QHBMTest(tf.test.TestCase):
     for q in raw_qubits:
       u += cirq.X(q)**s
   name = "TestQHBM"
-  raw_bit_circuit, raw_bit_symbols = qnn.build_bit_circuit(raw_qubits)
+  raw_bit_circuit = qnn.bit_circuit(raw_qubits)
+  raw_bit_symbols = list(sorted(tfq.util.get_circuit_symbols(raw_bit_circuit)))
   bit_symbols = tf.constant([str(s) for s in raw_bit_symbols])
 
   def test_init(self):
     """Confirms QHBM is initialized correctly."""
     test_ebm = ebm.Bernoulli(self.num_bits)
-    test_qnn = qnn.QNN(self.u, self.raw_phis_symbols)
+    test_qnn = qnn.QNN(self.u)
     test_qhbm = qhbm.QHBM(test_ebm, test_qnn, self.name)
     self.assertEqual(self.name, test_qhbm.name)
     self.assertEqual(test_ebm, test_qhbm.ebm)
@@ -68,7 +70,7 @@ class QHBMTest(tf.test.TestCase):
   def test_copy(self):
     """Confirms copy works correctly."""
     test_ebm = ebm.Bernoulli(self.num_bits)
-    test_qnn = qnn.QNN(self.u, self.raw_phis_symbols)
+    test_qnn = qnn.QNN(self.u)
     test_qhbm = qhbm.QHBM(test_ebm, test_qnn, self.name)
     qhbm_copy = test_qhbm.copy()
     self.assertEqual(test_qhbm.name, qhbm_copy.name)
@@ -101,7 +103,7 @@ def get_basic_qhbm(analytic=False):
   for s in phis_symbols:
     for q in qubits:
       u += cirq.X(q)**s
-  test_qnn = qnn.QNN(u, phis_symbols, analytic=analytic)
+  test_qnn = qnn.QNN(u, analytic=analytic)
   test_qnn._values.assign(initial_phis)
   name = "static_qhbm"
   return qhbm.QHBM(test_ebm, test_qnn, name=name)
@@ -181,13 +183,14 @@ class QHBMBasicFunctionTest(tf.test.TestCase):
     n_samples_0 = int(1e4)
     n_samples_1 = int(2e4)
     counts = tf.constant([n_samples_0, n_samples_1])
-    ragged_samples = test_qhbm.qnn.pulled_back_sample(circuits, counts)
+    ragged_samples = test_qhbm.qnn.pulled_back_sample(
+        circuits, counts, unique=False)
     test_samples_0 = ragged_samples[0].to_tensor()
     test_samples_1 = ragged_samples[1].to_tensor()
     self.assertEqual(n_samples_0, test_samples_0.shape[0])
     self.assertEqual(n_samples_1, test_samples_1.shape[0])
-    uniques_0, _ = ebm.unique_bitstrings_with_counts(test_samples_0)
-    uniques_1, _ = ebm.unique_bitstrings_with_counts(test_samples_1)
+    uniques_0, _ = util.unique_bitstrings_with_counts(test_samples_0)
+    uniques_1, _ = util.unique_bitstrings_with_counts(test_samples_1)
     self.assertEqual(1, uniques_0.shape[0])
     self.assertEqual(1, uniques_1.shape[0])
     self.assertAllEqual(tf.constant([0, 1, 0], dtype=tf.int8), uniques_0[0])
@@ -316,7 +319,7 @@ class ExactQHBMBasicFunctionTest(tf.test.TestCase):
                                            dtype=tf.float32))  # pin at |00>
     qubits = cirq.GridQubit.rect(1, n_bits)
     test_u = cirq.Circuit([cirq.H(qubits[0]), cirq.CNOT(qubits[0], qubits[1])])
-    test_qnn = qnn.QNN(test_u, [], analytic=True)
+    test_qnn = qnn.QNN(test_u, analytic=True)
     test_qhbm = qhbm.QHBM(test_ebm, test_qnn, name="bell")
     expected_dm = tf.constant(
         [[0.5, 0, 0, 0.5], [0, 0, 0, 0], [0, 0, 0, 0], [0.5, 0, 0, 0.5]],
